+++ module/Core/src/Crawling/CrawlingHelperInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Shlinkio\Shlink\Core\Crawling;
+
+interface CrawlingHelperInterface
+{
+    /**
+     * @return string[]|iterable
+     */
+    public function listCrawlableShortCodes(): iterable;
+}